Add m68k target for xosera_tester_fast
