+++ xosera_tester_fast/.gitignore
@@ -1,0 +1,7 @@
+temp/*
+*.o
+*.dis
+*.elf
+*.map
+*.sym
+
