+++ xosera_tester_fast/m68k/Makefile
@@ -1,0 +1,116 @@
+# Make Xosera tester for m68k target
+#
+# Copyright (c) 2021 Ross Bamford
+# MIT LICENSE
+
+EXTRA_CFLAGS?=-g -O3
+EXTRA_CXXFLAGS?=-g -O3
+SYSINCDIR?=../../../rosco_m68k/code/software/libs/build/include
+SYSLIBDIR?=../../../rosco_m68k/code/software/libs/build/lib
+DEFINES=-DROSCO_M68K
+CFLAGS=-std=c11 -ffreestanding -ffunction-sections -fdata-sections 					\
+				-Wall -Wextra -Werror -Wno-unused-function -pedantic 								\
+				-I$(SYSINCDIR) -I../include -I../include/arch/m68k	 								\
+				-mcpu=68010 -march=68010 -mtune=68010 $(DEFINES)
+CXXFLAGS=-std=c++17 -ffreestanding -ffunction-sections -fdata-sections 			\
+		  		-Wall -Wextra -Werror -Wno-unused-function -pedantic 							\
+			  	-I$(SYSINCDIR) -I../include -I../include/arch/m68k								\
+				  -mcpu=68010 -march=68010 -mtune=68010 $(DEFINES)
+GCC_LIBS=$(shell $(CC) --print-search-dirs 																	\
+		| grep libraries:\ =																										\
+    | sed 's/libraries: =/-L/g' 																						\
+    | sed 's/:/m68000\/ -L/g')m68000/
+LIBS=-lprintf -lcstdlib -lmachine -lstart_serial -ldebug_stub -lgcc
+ASFLAGS=-mcpu=68010 -march=68010
+LDFLAGS=-T $(SYSLIBDIR)/ld/serial/rosco_m68k_program.ld -L $(SYSLIBDIR) 		\
+				-Map=$(MAP) --gc-sections --oformat=elf32-m68k
+VASMFLAGS=-Felf -m68010 -quiet -Lnf $(DEFINES)
+CC=m68k-elf-gcc
+CXX=m68k-elf-g++
+AS=m68k-elf-as
+LD=m68k-elf-ld
+NM=m68k-elf-nm
+LD=m68k-elf-ld
+OBJDUMP=m68k-elf-objdump
+OBJCOPY=m68k-elf-objcopy
+SIZE=m68k-elf-size
+VASM=vasmm68k_mot
+RM=rm -f
+KERMIT=kermit
+SERIAL?=/dev/modem
+BAUD?=9600
+
+# Output config (assume name of directory)
+PROGRAM_BASENAME=xosera_tester_m68k
+
+# Set other output files using output basname
+ELF=$(PROGRAM_BASENAME).elf
+BINARY=$(PROGRAM_BASENAME).bin
+DISASM=$(PROGRAM_BASENAME).dis
+MAP=$(PROGRAM_BASENAME).map
+SYM=$(PROGRAM_BASENAME).sym
+
+# Assume source files in Makefile directory are source files for project
+CSOURCES=$(wildcard *.c)
+SSOURCES=$(wildcard *.S)
+ASMSOURCES=$(wildcard *.asm)
+SOURCES=$(CSOURCES) $(SSOURCES) $(ASMSOURCES)
+
+# Assume each source files makes an object file
+OBJECTS=$(addsuffix .o,$(basename $(SOURCES))) xosera_tests.o
+
+all: $(BINARY) $(DISASM)
+
+$(ELF) : $(OBJECTS)
+	$(LD) $(LDFLAGS) $(GCC_LIBS) $^ -o $@ $(LIBS)
+	$(NM) --numeric-sort $@ >$(SYM)
+	$(SIZE) $@
+	chmod a-x $@
+
+$(BINARY) : $(ELF)
+	$(OBJCOPY) -O binary $(ELF) $(BINARY)
+
+$(DISASM) : $(ELF)
+	$(OBJDUMP) --disassemble -S $(ELF) >$(DISASM)
+
+$(OBJECTS): Makefile
+
+xosera_tests.o: ../xosera_tests.cpp
+	$(CXX) -c $(CXXFLAGS) $(EXTRA_CXXFLAGS) -o $@ $<
+
+%.o : %.c
+	$(CC) -c $(CFLAGS) $(EXTRA_CFLAGS) -o $@ $<
+
+%.o : %.asm
+	$(VASM) $(VASMFLAGS) $(EXTRA_VASMFLAGS) -L $(basename $@).lst -o $@ $<
+
+# remove targets that can be generated by this Makefile
+clean:
+	$(RM) $(OBJECTS) $(ELF) $(BINARY) $(MAP) $(SYM) $(DISASM) $(addsuffix .lst,$(basename $(SSOURCES) $(ASMSOURCES)))
+
+disasm: $(DISASM)
+
+# hexdump of program binary
+dump: $(BINARY)
+	hexdump -C $(BINARY)
+
+# upload binary to rosco (if ready and kermit present)
+load: $(BINARY)
+	kermit -i -l $(SERIAL) -b $(BAUD) -s $(BINARY)
+
+# This is handy to test on Ubuntu
+test: rle_images.h $(BINARY) $(DISASM)
+	-killall screen && sleep 1
+	kermit -i -l $(SERIAL) -b $(BAUD) -s $(BINARY)
+	gnome-terminal --geometry=80x25 --title="rosco_m68k $(SERIAL)" -- screen $(SERIAL) $(BAUD)
+
+# This is handy to test on MacOS
+mactest: $(BINARY) $(DISASM)
+	-killall screen && sleep 1
+	kermit -i -l $(SERIAL) -b $(BAUD) -s $(BINARY)
+	echo "/usr/bin/screen $(SERIAL) $(BAUD)" > $(TMPDIR)/rosco_screen.sh
+	chmod +x $(TMPDIR)/rosco_screen.sh
+	open -b com.apple.terminal $(TMPDIR)/rosco_screen.sh
+
+# Makefile magic (for "phony" targets that are not real files)
+.PHONY: all clean dump disasm load test mactest
